a2a/parallel-agents/pyproject.toml
[project]
name = "parallel-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.10.0",
]
